requirements.txt
click==8.1.3
Flask==2.2.2
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
protobuf==3.20.3
python-dotenv==0.21.1
waitress==2.1.2
Werkzeug==2.2.2
zipp==3.12.0
